Separate component and system into different files
